Fix type conversion for parameters passed by reference (gh#yast/ycp-killer#486)
https://github.com/yast/ycp-killer/pull/486

When returning from a Perl function, if some arguments were passed
by reference we must copy them across the language barrier,
and use the correct YCP type to convert the weakly typed Perl
values.

The code used to look at the type of the referenced YCP variable,
which is wrong in case it is actually a proxy for a Ruby variable:
a parameter of type list<integer>& pointed to a list<integer> YCP
variable but only to a list<any> Ruby variable. Therefore the value
from Perl came out as list<string>.

The fix is to use the formal parameters of the function.